requirements.txt
streamlit>=1.30.0
openai>=1.30.1
redis>=5.0.0
python-dotenv>=1.0.1